requirements.txt
together
python-dotenv
pyside6